backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = "An AI-powered platform to assist students with their academic needs."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "firebase-admin>=7.1.0",
    "httpx>=0.28.1",
    "jose>=1.0.0",
    "pydantic[email]>=2.12.0",
    "pyrebase4>=4.8.0",
    "pytest>=9.0.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "httpx>=0.28.1",
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.6",
]

[tool.pytest]
minversion = "9.0"
addopts = ["-ra", "-q"]
testpaths = [
    "tests",
]

